A7
Changed API endpoint to a new, better, faster endpoint. Also reduced the minimum update time from 5 minutes to 1 minute as there's no possibility of overloading the endpoint.
Also when creating a gameday having failing checks no longer blocks the creation of the gameday, this is only to be able to get a gameday lined up to release it when ready.